+++ web/src/classes/ImportedBeer.ts
@@ -1,0 +1,19 @@
+import { IImportedBeer } from '@/interfaces/IImportedBeer';
+import { IMashStep } from '@/interfaces/IMashStep';
+import { INotification } from '@/interfaces/INotification';
+
+export default class ImportedBeer implements IImportedBeer {
+  public name: string | null = null;
+
+  public mashSteps: Array<IMashStep> = [];
+
+  public mashNotifications: Array<INotification> = [];
+
+  public mashNotificationsGrouped: Array<INotification> = [];
+
+  public boilSteps: Array<IMashStep> = [];
+
+  public boilNotifications: Array<INotification> = [];
+
+  public boilNotificationsGrouped: Array<INotification> = [];
+}