+++ web/src/interfaces/IPidSettings.ts
@@ -2,5 +2,9 @@ export interface IPidSettings {
   kP: number;
   kI: number;
   kD: number;
+  boilkP: number;
+  boilkI: number;
+  boilkD: number;
   pidLoopTime: number;
+  stepInterval: number;
 }