+++ web/src/views/PidSettings.vue
@@ -1,5 +1,6 @@
 <script lang="ts" setup>
 import { inject, onBeforeUnmount, onMounted, ref } from 'vue';
+import { mdiHelp } from '@mdi/js';
 import WebConn from '@/helpers/webConn';
 import { IPidSettings } from '@/interfaces/IPidSettings';
 
@@ -9,7 +10,11 @@ const pidSettings = ref<IPidSettings>({ // add default value, vue heeft issues m
   kP: 0,
   kI: 0,
   kD: 0,
+  boilkP: 0,
+  boilkI: 0,
+  boilkD: 0,
   pidLoopTime: 60,
+  stepInterval: 60,
 });
 
 const getData = async () => {
@@ -59,24 +64,64 @@ const save = async () => {
 
       <v-row>
         <v-col cols="12" md="3">
-          <v-text-field type="number" v-model.number="pidSettings.kP" label="P" />
+          <v-text-field type="number" v-model.number="pidSettings.pidLoopTime" label="PID Loop Time">
+            <template v-slot:append>
+              <v-tooltip text="The time in seconds between PID calculations, since water heating is a slow process this also works best when slow ex. 60sec">
+                <template v-slot:activator="{ props }">
+                  <v-icon size="small" v-bind="props">{{mdiHelp}}</v-icon>
+                </template>
+              </v-tooltip>
+            </template>
+          </v-text-field>
         </v-col>
       </v-row>
+
       <v-row>
+        <v-col cols="12" md="3">
+          <v-text-field type="number" v-model.number="pidSettings.stepInterval" label="Steps every x seconds">
+            <template v-slot:append>
+              <v-tooltip text="The time between steps less means more substeps will be added">
+                <template v-slot:activator="{ props }">
+                  <v-icon size="small" v-bind="props">{{mdiHelp}}</v-icon>
+                </template>
+              </v-tooltip>
+            </template>
+          </v-text-field>
+        </v-col>
+      </v-row>
+
+      <div class="text-subtitle-2 mt-4 mb-2">Mash</div>
+
+      <v-divider :thickness="7" />
+
+      <v-row class="mt-4 mb-2">
+        <v-col cols="12" md="3">
+          <v-text-field type="number" v-model.number="pidSettings.kP" label="P" />
+        </v-col>
         <v-col cols="12" md="3">
           <v-text-field type="number" v-model.number="pidSettings.kI" label="I" />
         </v-col>
-      </v-row>
-      <v-row>
         <v-col cols="12" md="3">
           <v-text-field type="number" v-model.number="pidSettings.kD" label="D" />
         </v-col>
       </v-row>
-      <v-row>
+
+      <div class="text-subtitle-2 mt-4 mb-2">Boil</div>
+
+      <v-divider :thickness="7" />
+
+      <v-row class="mt-4 mb-2">
         <v-col cols="12" md="3">
-          <v-text-field type="number" v-model.number="pidSettings.pidLoopTime" label="PID Loop Time" />
+          <v-text-field type="number" v-model.number="pidSettings.boilkP" label="P" />
+        </v-col>
+        <v-col cols="12" md="3">
+          <v-text-field type="number" v-model.number="pidSettings.boilkI" label="I" />
+        </v-col>
+        <v-col cols="12" md="3">
+          <v-text-field type="number" v-model.number="pidSettings.boilkD" label="D" />
         </v-col>
       </v-row>
+
       <v-row>
         <v-col cols="12" md="3">
           <v-btn color="success" class="mt-4 mr-2" @click="save"> Save </v-btn>