+++ Changelog.md
@@ -1,3 +1,11 @@
+# Version 1.4.0
+- Extra setting to configure stepInterval.
+- Optimized substeps calculation.
+- Optimized control loop, trigger pid faster after target changes.
+- Different PID settings for Boil/Mash.
+- Import of BeerXML.
+- Dark/Light mode via clientsettings.
+
 # Version 1.3.1
 - Bugfix: One cannot delete steps.
 - Bugfix: Wifi settings one phone responsiveness is wrong.